[FE][Refactor] #13: Content 분리 및 폴더,파일명 변경
