+++ frontend/src/component/content/Content.tsx
@@ -4,18 +4,9 @@ import { MdGroup, MdMoreVert } from 'react-icons/md';
 interface IContentProps {
   title: string;
   time: string;
-  person?: string;
+  person: number;
   link: string;
 }
-
-/**
- * `Content` 컴포넌트는 경로의 제목, 시간, 인원 수를 표시하며, 드롭다운 메뉴를 통해 수정 및 삭제 기능을 제공합니다.
- *
- * @param title 제목
- * @param time 시간
- * @param person 인원 수
- * @returns JSX.Element
- */
 
 export const Content = (props: IContentProps) => {
   return (
@@ -28,10 +19,10 @@ export const Content = (props: IContentProps) => {
         <div className="flex items-center text-sm leading-5 text-gray-500">
           <span className="mr-4">시간</span>
           <span className="mr-6">{props.time}</span>
-          {props.person && (
+          {props.person > 0 && (
             <>
               <MdGroup className="mr-2 h-5 w-5" />
-              <span>{props.person}</span>
+              <span>{props.person}명</span>
             </>
           )}
         </div>
