+++ frontend/src/stories/Content.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { Content } from '@/component/BottomSheet/Content';
+import { Content } from '@/component/content/Content';
 
 const meta = {
   title: 'Components/Content',
@@ -11,7 +11,7 @@ const meta = {
   argTypes: {
     title: { control: 'text', description: 'channel의 제목을 나타낸다.' },
     time: { control: 'text', description: '출발지부터 목적지까지 소요시간을 나타낸다.' },
-    person: { control: 'text', description: 'channel에 접속한 인원 수를 나타냅니다.' },
+    person: { control: 'number', description: 'channel에 접속한 인원 수를 나타냅니다.' },
   },
 } satisfies Meta<typeof Content>;
 
@@ -22,7 +22,8 @@ export const Default: Story = {
   args: {
     title: '아들네 집으로',
     time: '0시간 30분',
-    person: '2명',
+    person: 2,
+    link: '/Default',
   },
 };
 
@@ -30,7 +31,8 @@ export const NoPerson: Story = {
   args: {
     title: '친구 집 방문',
     time: '1시간 20분',
-    person: '',
+    person: 0,
+    link: '/NoPerson',
   },
 };
 
@@ -38,6 +40,7 @@ export const LongText: Story = {
   args: {
     title: '대한민국 경기도 성남시 분당구 네이버 본사까지 가는 경로 안내 채널',
     time: '2시간 45분',
-    person: '4명',
+    person: 4,
+    link: '/LongText',
   },
 };
